Code cleanup
- removed unused varialbles
- eliminated all warnings when compiling with avr-gcc4.7.2
- fixed signed/unsigned mismatchs
- moved public variables to proper place

+++ Marlin/planner.cpp
@@ -194,7 +194,7 @@ void calculate_trapezoid_for_block(block_t *block, float entry_factor, float exi
     accelerate_steps = ceil(
     intersection_distance(block->initial_rate, block->final_rate, acceleration, block->step_event_count));
     accelerate_steps = max(accelerate_steps,0); // Check limits due to numerical round-off
-    accelerate_steps = min(accelerate_steps,block->step_event_count);
+    accelerate_steps = min((unsigned long)accelerate_steps,block->step_event_count);
     plateau_steps = 0;
   }
